requirements.txt
attrs==23.2.0
blinker==1.6.2
boto3==1.34.19
botocore==1.34.22
cement==2.8.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.3
cryptography==42.0.5
dnspython==2.4.2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-CKEditor==0.5.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.3
pathspec==0.10.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==4.6.0
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.10.0
semantic-version==2.8.5
service-identity==24.1.0
six==1.16.0
termcolor==1.1.0
urllib3==1.26.16
wcwidth==0.1.9
Werkzeug==2.3.7
wtf==0.1
WTForms==3.1.2
